split views and api endpoints
